+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFX4XN6J147EY2E48Q8286CY"
+        "01KFZQC147W17BJ2EQE0RJQES9"
     )
     r"""The unique identifier of the data source"""