+++ packages/orq-rc/RELEASES.md
@@ -5059,3 +5059,13 @@ Based on:
 - [python v4.3.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.3 - packages/orq-rc
+
+## 2026-01-27 12:40:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.698.0 (2.801.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.5 - packages/orq-rc